refactor: move file package to pkg directory
Relocate file package from root to pkg/file/ as part of
framework reorganization. No functional changes.
