+++ pkg/file/file_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/dmitrymomot/saaskit/file"
+	"github.com/dmitrymomot/saaskit/pkg/file"
 )
 
 func createFileHeader(filename string, content []byte) *multipart.FileHeader {
